+++ changes.md
@@ -3,15 +3,17 @@
 ## Beta 0.2.1
 This patch fixes a number of issues related to UI.
 * Cards
-    * Location now properly announced (currently for shop)
+    * Location now properly announced (currently excluding compendium)
+* Combat Rewards Screen
+    * All rewards now have location information and properly update buffers
 * Monsters
     * Power tips (the descriptions of any applied powers in the buffer) will no longer incorrectly show formatting information if that information occurs at the start of the first word of the effect description (for example #y at the beginning of weak's description)
 * Potions
-    * Potion location is now properly announced (currently for main screen, shop, and combat rewards)
+    * Potion location is now properly announced (excluding compendium)
     * Fixed a bug where you could not read the tooltip for a nempty potion slot
     * Potion rarity has been added to the potion buffer
 * Relics
-    * Location is now properly announced (now for main screen, rewards, and shops)
+    * Location is now properly announced (excluding compendium)
 * Shops
     * All shop items now have location information approximately corresponding to their on screen position
     * The button to sell a card is now actually announced as a button and has location information